ref(node): Refactor orchestrion config into separate files
