+++ packages/server-utils/src/orchestrion/channels.ts
@@ -1,3 +1,11 @@
+import { mysqlChannels } from './config/mysql';
+import { lruMemoizerChannels } from './config/lru-memoizer';
+import { ioredisChannels } from './config/ioredis';
+import { pgChannels } from './config/pg';
+import { openaiChannels } from './config/openai';
+import { anthropicAiChannels } from './config/anthropic-ai';
+import { vercelAiChannels } from './config/vercel-ai';
+
 /**
  * Fully-qualified `diagnostics_channel` names that orchestrion publishes to.
  *
@@ -12,33 +20,13 @@
  * they don't drift apart and silently stop firing.
  */
 export const CHANNELS = {
-  MYSQL_QUERY: 'orchestrion:mysql:query',
-  LRU_MEMOIZER_LOAD: 'orchestrion:lru-memoizer:load',
-  IOREDIS_COMMAND: 'orchestrion:ioredis:command',
-  IOREDIS_CONNECT: 'orchestrion:ioredis:connect',
-  PG_QUERY: 'orchestrion:pg:query',
-  PG_CONNECT: 'orchestrion:pg:connect',
-  PGPOOL_CONNECT: 'orchestrion:pg-pool:connect',
-  OPENAI_CHAT: 'orchestrion:openai:chat',
-  OPENAI_RESPONSES: 'orchestrion:openai:responses',
-  OPENAI_EMBEDDINGS: 'orchestrion:openai:embeddings',
-  OPENAI_CONVERSATIONS: 'orchestrion:openai:conversations',
-  ANTHROPIC_CHAT: 'orchestrion:@anthropic-ai/sdk:chat',
-  ANTHROPIC_MODELS: 'orchestrion:@anthropic-ai/sdk:models',
-  ANTHROPIC_MESSAGES_STREAM: 'orchestrion:@anthropic-ai/sdk:messages-stream',
-  // Vercel AI (`ai`) v6: orchestrion injects these so the same channel-based
-  // integration that consumes `ai`'s native `ai:telemetry` channel (v7) can
-  // also instrument v6. Each maps to a top-level function in `ai`'s bundle.
-  VERCEL_AI_GENERATE_TEXT: 'orchestrion:ai:generateText',
-  VERCEL_AI_STREAM_TEXT: 'orchestrion:ai:streamText',
-  VERCEL_AI_EMBED: 'orchestrion:ai:embed',
-  VERCEL_AI_EMBED_MANY: 'orchestrion:ai:embedMany',
-  VERCEL_AI_EXECUTE_TOOL_CALL: 'orchestrion:ai:executeToolCall',
-  // `resolveLanguageModel` is the single chokepoint every model call flows
-  // through; we wrap it to monkey-patch `doGenerate`/`doStream` on the returned
-  // model (the model-call site itself is an inline call with no injectable
-  // definition).
-  VERCEL_AI_RESOLVE_LANGUAGE_MODEL: 'orchestrion:ai:resolveLanguageModel',
+  ...mysqlChannels,
+  ...lruMemoizerChannels,
+  ...ioredisChannels,
+  ...pgChannels,
+  ...openaiChannels,
+  ...anthropicAiChannels,
+  ...vercelAiChannels,
 } as const;
 
 export type ChannelName = (typeof CHANNELS)[keyof typeof CHANNELS];
