+++ packages/server-utils/src/orchestrion/config/mysql.ts
@@ -1,0 +1,13 @@
+import type { InstrumentationConfig } from '@apm-js-collab/code-transformer';
+
+export const mysqlConfig = [
+  {
+    channelName: 'query',
+    module: { name: 'mysql', versionRange: '>=2.0.0 <3', filePath: 'lib/Connection.js' },
+    functionQuery: { expressionName: 'query', kind: 'Auto' },
+  },
+] satisfies InstrumentationConfig[];
+
+export const mysqlChannels = {
+  MYSQL_QUERY: 'orchestrion:mysql:query',
+} as const;
